Address pipeline cleanup review feedback
